refactor using args
