+++ fastcoref/__init__.py
@@ -1,1 +1,1 @@
-from .modeling import FCoref, LingMessCoref
+from .modeling import FCoref, LingMessCoref, CorefArgs, CorefResult
