+++ tests/local_test.py
@@ -1,7 +1,14 @@
-from fastcoref import FCoref, LingMessCoref
+from fastcoref import FCoref, LingMessCoref, CorefArgs
 
 texts = ['We are so happy to see you using our coref package. This package is very fast!',
          'The man tried to put the boot on his foot but it was too small.']
+
+
+# args = CorefArgs(
+#     model_name_or_path='you-awesome-model',
+#     device='cpu'
+# )
+# model = FCoref(args=args)
 
 
 model = FCoref()
@@ -25,6 +32,10 @@ print(preds[0].get_logit(span_i=(33, 50), span_j=(52, 64)))
 print(preds[1].get_clusters())
 print(preds[1].get_clusters(as_strings=False))
 print(preds[1].get_logit(span_i=(21, 29), span_j=(46, 48)))
+
+
+
+
