+++ tests/test_fcoref.py
@@ -1,20 +1,23 @@
 import unittest
 
-from fastcoref.modeling import CorefResult
-
-from fastcoref import FCoref
+from fastcoref import FCoref, CorefArgs, CorefResult
 
 
 class TestFCoref(unittest.TestCase):
     def setUp(self) -> None:
         self.test_text = ['We are so happy to see you using our coref package. This package is very fast!',
                           'The man tried to put the boot on his foot but it was too small.']
-        self.model = FCoref()
-        device = 'cpu'
+        args = CorefArgs(
+            model_name_or_path='biu-nlp/f-coref',
+            device='cpu',
+            max_tokens_in_batch=7000
+        )
+        self.model = FCoref(args=args)
+
         self.expected_clusters = [[[(0, 2), (33, 36)], [(33, 50), (52, 64)]],
-                                  [[(0, 7), (33, 36)], [(33, 41), (46, 48)]]]
+                                  [[(0, 7), (33, 36)], [(21, 29), (46, 48)]]]
         self.expected_clusters_strings = [[['We', 'our'], ['our coref package', 'This package']],
-                                          [['The man', 'his'], ['his foot', 'it']]]
+                                          [['The man', 'his'], ['the boot', 'it']]]
 
     def test_predict_with_unexpected_object(self):
         texts = {'text1': 'sss'}
@@ -51,12 +54,9 @@ class TestFCoref(unittest.TestCase):
             self.assertIsInstance(res_obj, CorefResult)
             self.assertListEqual(res_obj.get_clusters(as_strings=False), self.expected_clusters[i])
 
-
     def test_get_logits(self):
         preds = self.model.predict(texts=self.test_text)
         self.assertIsInstance(preds, list)
 
         self.assertGreater(preds[0].get_logit(span_i=(33, 50), span_j=(52, 64)), 0)
-        with self.assertRaises(ValueError) as exc:
-            logit = preds[1].get_logit(span_i=(21, 29), span_j=(46, 48))
-        self.assertEqual(str(exc.exception), f'span_i="the boot" is not an entity in this model!')
+        self.assertGreater(preds[1].get_logit(span_i=(21, 29), span_j=(46, 48)), 0)
